refactor eval out of lib.rs
